pyproject.toml
[project]
name = "patient-readmission-risk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beaverfe>=0.2.1",
    "evidently>=0.7.11",
    "fastapi>=0.116.1",
    "numpy>=2.3.1",
    "optuna>=4.4.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "wandb>=0.21.0",
    "xgboost>=3.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
]
commits = [
    "pre-commit>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
